+++ engine/api/link.go
@@ -2,10 +2,13 @@ package api
 
 import (
 	"context"
+
 	"net/http"
+	"time"
 
 	"github.com/gorilla/mux"
 
+	"github.com/ovh/cds/engine/api/link"
 	"github.com/ovh/cds/engine/service"
 	"github.com/ovh/cds/sdk"
 )
@@ -54,5 +57,73 @@ func (api *API) postAskLinkExternalUserWithCDSHandler() service.Handler {
 			return err
 		}
 		return service.WriteJSON(w, redirect, http.StatusOK)
+	}
+}
+
+func (api *API) postLinkExternalUserWithCDSHandler() service.Handler {
+	return func(ctx context.Context, w http.ResponseWriter, r *http.Request) error {
+		vars := mux.Vars(r)
+
+		u := getUserConsumer(ctx)
+
+		// Extract consumer type from request, is invalid or not in api drivers list return an error
+		consumerType := sdk.AuthConsumerType(vars["consumerType"])
+		linkDriver, has := api.LinkDrivers[consumerType]
+		if !has {
+			return sdk.WithStack(sdk.ErrInvalidData)
+		}
+
+		// Extract and validate signin request
+		var req sdk.AuthConsumerSigninRequest
+		if err := service.UnmarshalBody(r, &req); err != nil {
+			return err
+		}
+
+		// Extract and validate signin state
+		switch x := linkDriver.GetDriver().(type) {
+		case sdk.DriverWithSignInRequest:
+			if err := x.CheckSigninRequest(req); err != nil {
+				return err
+			}
+		default:
+			return sdk.WithStack(sdk.ErrInvalidData)
+		}
+		switch x := linkDriver.GetDriver().(type) {
+		case sdk.DriverWithSigninStateToken:
+			if err := x.CheckSigninStateToken(req); err != nil {
+				return err
+			}
+		default:
+			sdk.WithStack(sdk.ErrInvalidData)
+		}
+
+		// Convert code to external user info
+		userInfo, err := linkDriver.GetUserInfo(ctx, req)
+		if err != nil {
+			return err
+		}
+
+		tx, err := api.mustDB().Begin()
+		if err != nil {
+			return sdk.WithStack(err)
+		}
+		defer tx.Rollback() // nolint
+
+		userLink := sdk.UserLink{
+			AuthentifiedUserID: u.AuthConsumerUser.AuthentifiedUserID,
+			Created:            time.Now(),
+			Username:           userInfo.Username,
+			ExternalID:         userInfo.ExternalID,
+			Type:               string(consumerType),
+		}
+		if err := link.Insert(ctx, tx, &userLink); err != nil {
+			return err
+		}
+
+		if err := tx.Commit(); err != nil {
+			return sdk.WithStack(err)
+		}
+
+		return service.WriteJSON(w, userLink, http.StatusOK)
 	}
 }
